# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SIGET — API de Gestión de Maquinaria y Operaciones
  description: |
    API REST para la gestión de operaciones diarias de la flota vehicular, autenticación de usuarios
    y parámetros generales del sistema.

    ## Autenticación

    La mayoría de los endpoints requieren un token **JWT**. Flujo recomendado:

    1. Llamar a `POST /auth/login` con `email` y `password` para obtener un `access_token`.
    2. Enviar dicho token en la cabecera `Authorization: Bearer {access_token}` en cada petición protegida.
    3. Cuando el token esté por expirar, llamar a `POST /auth/refresh` (con el token actual todavía vigente)
       para obtener uno nuevo.
    4. Llamar a `POST /auth/logout` para invalidar el token de forma explícita.

    ## Formato de errores

    - **422 Unprocessable Content**: error de validación. Responde `{ "success": false, "message": "...", "errors": { "campo": ["mensaje"] } }`.
    - **401 Unauthorized**: token ausente, inválido o expirado.
    - **403 Forbidden**: el usuario está autenticado pero no tiene permiso sobre el recurso solicitado.
    - **404 Not Found**: el recurso indicado en la ruta no existe.
    - **500 Internal Server Error**: error inesperado del servidor. Responde `{ "message": "...", "error": "..." }`.

    ## Usuarios de prueba

    Estas cuentas están precargadas en el entorno de pruebas y sirven para probar la
    API con cada rol desde `POST /auth/login`:

...
  version: "1.9.0"
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /ordenes-trabajo/{orden}/iniciar:
    patch:
      tags: [Órdenes de Trabajo]
      summary: Marcar el mantenimiento como iniciado
      description: |
        El técnico marca el inicio del trabajo: la orden pasa de `PENDIENTE` a `EN_EJECUCION`
        y se registra la fecha/hora de ejecución. A partir de aquí puede ir agregando ítems al
        detalle (`POST /ordenes-trabajo/{orden}/detalles`) y, al terminar, culminarla.

        Sin cuerpo. Sólo funciona si la orden está `PENDIENTE`; en cualquier otro estado
        responde `422`. Misma restricción de acceso que el resto de acciones del técnico.
      parameters:
        - name: orden
          in: path
          required: true
          description: ID de la orden de trabajo.
          schema: { type: integer }
          example: 1
      responses:
        '200':
          description: Mantenimiento iniciado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Mantenimiento iniciado. }
                  data: { $ref: '#/components/schemas/OrdenTrabajo' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El técnico autenticado intenta iniciar una orden que no tiene asignada.
          content:
            application/json:
              example:
                message: No tiene acceso a esta orden de trabajo.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: La orden no está `PENDIENTE`.
          content:
            application/json:
              example:
                message: Sólo se puede iniciar una orden que está PENDIENTE.
  /ordenes-trabajo/{orden}/culminar:
    post:
      tags: [Órdenes de Trabajo]
      summary: Marcar el mantenimiento como culminado
      description: |
        El técnico cierra la ejecución: la orden pasa de `EN_EJECUCION` a `CULMINADO`, se
        registran las **lecturas finales** del vehículo y una observación opcional, y el
        detalle de trabajo queda **congelado** (ya no admite más ítems ni ediciones).

        Requisitos:
        - La orden debe estar `EN_EJECUCION` (debe haberse iniciado antes); en otro estado
          responde `422`.
        - Debe tener **al menos un ítem** de detalle registrado; si no, responde `422`.

        El campo de lectura final obligatorio depende del `tipo_medicion` del vehículo:
        `kilometraje_actual` si mide por kilometraje, `horometro_actual` si mide por horómetro.

        Misma restricción de acceso que el resto de acciones del técnico.
      parameters:
        - name: orden
          in: path
          required: true
          description: ID de la orden de trabajo.
          schema: { type: integer }
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdenTrabajoCulminarRequest'
            example:
              kilometraje_actual: 85300
              observacion: Se cambió aceite y filtros. Todo en orden.
      responses:
        '200':
          description: Mantenimiento culminado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Mantenimiento culminado exitosamente. }
                  data: { $ref: '#/components/schemas/OrdenTrabajo' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El técnico autenticado intenta culminar una orden que no tiene asignada.
          content:
            application/json:
              example:
                message: Sólo puede culminar la ejecución de las órdenes que tiene asignadas.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: |
            Error de validación, o la orden no está `EN_EJECUCION`, o no tiene ningún ítem de
            detalle registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultValidationErrorResponse'
              examples:
                validacion:
                  summary: Falta la lectura final
                  value:
                    success: false
                    message: Los datos enviados no son válidos.
                    errors:
                      kilometraje_actual: ["El kilometraje actual es obligatorio."]
                sin_iniciar:
                  summary: La orden no fue iniciada
                  value:
                    message: Sólo se puede culminar una orden que está EN_EJECUCION (primero debe iniciarla).
                sin_detalle:
                  summary: Sin ítems de detalle
                  value:
                    message: Debe registrar al menos un ítem del detalle antes de culminar la orden.
  /ordenes-trabajo/{orden}/detalles:
    post:
      tags: [Órdenes de Trabajo]
      summary: Registrar un ítem del detalle de trabajo
      description: |
        Registra **un** ítem del detalle de trabajo de la orden (un repuesto/insumo aplicado, o
        una acción de mano de obra sin repuesto asociado). El técnico llama a este endpoint cada
        vez que completa una acción: los ítems se agregan de a uno, no hace falta enviarlos
        todos juntos.

        Se envía como `application/json`. `id_orden_trabajo` se toma de la URL y no debe
        enviarse.

        **Restricción de acceso:** un jefe de área/administrador puede registrar el detalle de
        cualquier orden; un técnico de mantenimiento, sólo el de las órdenes que tiene asignadas
        (`id_usuario_ejecuta`). Cualquier otro caso responde `403`.

        El campo de lectura obligatorio (`horometro` **o** `kilometraje`) depende del
        `tipo_medicion` del vehículo de la orden. Sólo se puede registrar detalle mientras la
        orden está `PENDIENTE` o `EN_EJECUCION`; si ya fue culminada/verificada/cancelada
        responde `422`.
      parameters:
        - name: orden
          in: path
          required: true
          description: ID de la orden de trabajo.
          schema: { type: integer }
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetalleMantenimientoStoreRequest'
            example:
              id_tipo_mantenimiento: 3
              id_repuesto: 8
              fecha: "2026-08-27"
              kilometraje: 85230.5
              cantidad: 2
      responses:
        '201':
          description: Ítem de detalle registrado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleMantenimientoStoreResponse'
              example:
                message: Detalle de mantenimiento registrado exitosamente.
                data:
                  id: 15
                  id_orden_trabajo: 1
                  id_repuesto: 8
                  id_tipo_mantenimiento: 3
                  fecha: "2026-08-27"
                  horometro: null
                  kilometraje: "85230.50"
                  cantidad: 2
                  created_at: "2026-08-27T14:30:00.000000Z"
                  updated_at: "2026-08-27T14:30:00.000000Z"
                  repuesto:
                    id: 8
                    nombre_repuesto: Filtro de aceite
                    codigo_repuesto: FIL-001
                    unidad_medida: UNIDAD
                  tipo_mantenimiento:
                    id: 3
                    tipo_mantenimiento: Cambio de aceite
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El técnico autenticado intenta registrar detalle en una orden que no tiene asignada.
          content:
            application/json:
              example:
                message: Sólo puede registrar el detalle de las órdenes que tiene asignadas.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: |
            Error de validación, o la orden ya no admite cambios en su detalle (estado distinto
            de `PENDIENTE`/`EN_EJECUCION`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultValidationErrorResponse'
              examples:
                validacion:
                  summary: Error de validación
                  value:
                    success: false
                    message: Los datos enviados no son válidos.
                    errors:
                      id_tipo_mantenimiento: ["Debe indicar el tipo de mantenimiento."]
                      kilometraje: ["El kilometraje es obligatorio."]
                orden_congelada:
                  summary: La orden ya fue culminada
                  value:
                    message: "La orden ya no admite cambios en su detalle de trabajo (sólo mientras está PENDIENTE o EN_EJECUCION)."
  /ordenes-trabajo/{orden}/detalles/{detalle}:
    parameters:
      - name: orden
        in: path
        required: true
        description: ID de la orden de trabajo.
        schema: { type: integer }
        example: 1
      - name: detalle
        in: path
        required: true
        description: ID del ítem de detalle (debe pertenecer a la orden).
        schema: { type: integer }
        example: 15

    put:
      tags: [Órdenes de Trabajo]
      summary: Corregir un ítem del detalle de trabajo
      description: |
        Corrige un ítem del detalle ya registrado (p. ej. una `cantidad` o una lectura mal
        cargada). Acepta `PUT` o `PATCH`, pero **no es un parche parcial**: el cuerpo tiene los
        mismos campos y las mismas reglas que `POST /ordenes-trabajo/{orden}/detalles` y debe
        reenviarse completo.

        Sólo mientras la orden está `PENDIENTE` o `EN_EJECUCION`; misma restricción de acceso
        que el registro de detalle (un técnico, sólo el de sus órdenes asignadas).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetalleMantenimientoStoreRequest'
            example:
              id_tipo_mantenimiento: 3
              id_repuesto: 8
              fecha: "2026-08-27"
              kilometraje: 85235
              cantidad: 3
      responses:
        '200':
          description: Ítem de detalle actualizado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleMantenimientoStoreResponse'
              example:
                message: Detalle de mantenimiento actualizado exitosamente.
                data:
                  id: 15
                  id_orden_trabajo: 1
                  id_repuesto: 8
                  id_tipo_mantenimiento: 3
                  fecha: "2026-08-27"
                  horometro: null
                  kilometraje: "85235.00"
                  cantidad: 3
                  created_at: "2026-08-27T14:30:00.000000Z"
                  updated_at: "2026-08-27T15:10:00.000000Z"
                  repuesto: { id: 8, nombre_repuesto: Filtro de aceite, codigo_repuesto: FIL-001, unidad_medida: UNIDAD }
                  tipo_mantenimiento: { id: 3, tipo_mantenimiento: Cambio de aceite }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El técnico autenticado intenta editar el detalle de una orden que no tiene asignada.
          content:
            application/json:
              example:
                message: Sólo puede registrar el detalle de las órdenes que tiene asignadas.
        '404':
          description: La orden o el ítem de detalle no existen, o el ítem no pertenece a esa orden.
        '422':
          description: |
            Error de validación, o la orden ya no admite cambios en su detalle (estado distinto
            de `PENDIENTE`/`EN_EJECUCION`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultValidationErrorResponse'
              examples:
                validacion:
                  summary: Error de validación
                  value:
                    success: false
                    message: Los datos enviados no son válidos.
                    errors:
                      cantidad: ["La cantidad es obligatoria."]
                orden_congelada:
                  summary: La orden ya fue culminada
                  value:
                    message: "La orden ya no admite cambios en su detalle de trabajo (sólo mientras está PENDIENTE o EN_EJECUCION)."

    delete:
      tags: [Órdenes de Trabajo]
      summary: Eliminar un ítem del detalle de trabajo
      description: |
        Elimina un ítem del detalle registrado por error. Sólo mientras la orden está
        `PENDIENTE` o `EN_EJECUCION`; misma restricción de acceso que el registro de detalle.
      responses:
        '200':
          description: Ítem de detalle eliminado exitosamente.
          content:
            application/json:
              example:
                message: Detalle de mantenimiento eliminado exitosamente.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El técnico autenticado intenta eliminar detalle de una orden que no tiene asignada.
          content:
            application/json:
              example:
                message: No tiene acceso a esta orden de trabajo.
        '404':
          description: La orden o el ítem de detalle no existen, o el ítem no pertenece a esa orden.
        '422':
          description: La orden ya no admite cambios en su detalle (estado distinto de `PENDIENTE`/`EN_EJECUCION`).
          content:
            application/json:
              example:
                message: "La orden ya no admite cambios en su detalle de trabajo (sólo mientras está PENDIENTE o EN_EJECUCION)."

  # ──────────────────────────────────────────────────────────────────────
  # NOTIFICACIONES
  # ──────────────────────────────────────────────────────────────────────
  # ──────────────────────────────────────────────────────────────────────
  # VEHÍCULOS
  # ──────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido en `POST /auth/login`. Enviar como
        `Authorization: Bearer {access_token}`.
  responses:
    Unauthorized:
      description: No autenticado, token ausente/inválido/expirado.
      content:
        application/json:
          example:
            message: Unauthenticated.
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          example:
            message: "No query results for model [App\\Models\\OperacionDiaria] 999"
  schemas:
...
    DefaultValidationErrorResponse:
      type: object
      description: |
        Formato **estándar** de error de validación (sin personalizar). A diferencia de
        `ValidationErrorResponse` (usado por algunos endpoints, como el de operación diaria),
        aquí **no** se envuelve con la clave `success`, y el `message` de nivel superior queda
        fijo en inglés (`"The given data was invalid."`) sin traducir. Los mensajes dentro de
        `errors` sí están en español.
      properties:
        message: { type: string, example: "The given data was invalid." }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          example:
            campo: ["mensaje de error"]

    # ── Carga de Combustible ─────────────────────────────────────────────
...
    OrdenTrabajo:
      type: object
      description: Orden de trabajo (Paso 2/3 del flujo de mantenimiento) con su detalle de trabajo.
      properties:
        id: { type: integer, example: 1 }
        id_solicitud_mantenimiento: { type: integer, nullable: true, example: 12 }
        id_vehiculo: { type: integer, example: 1 }
        id_conductor: { type: integer, nullable: true, example: 1 }
        id_taller: { type: integer, nullable: true, description: "`null` en órdenes internas.", example: null }
        id_usuario_emite: { type: integer, example: 2 }
        id_usuario_ejecuta: { type: integer, description: Técnico de mantenimiento responsable de ejecutar la orden. Se asigna al emitirla., example: 5 }
        nro_orden: { type: integer, example: 1 }
        gestion: { type: string, example: "2026" }
        fecha_emision: { type: string, format: date-time }
        fecha_ejecucion: { type: string, format: date-time, nullable: true }
        fecha_culminacion: { type: string, format: date-time, nullable: true }
        nota_emisor: { type: string, nullable: true, example: Revisar también el sistema de frenos. }
        observacion: { type: string, nullable: true }
        tipo_mantenimiento: { type: string, enum: [PREVENTIVO, CORRECTIVO], example: PREVENTIVO }
        kilometraje_actual: { type: integer, nullable: true, example: 85000 }
        horometro_actual: { type: integer, nullable: true, example: null }
        estado_orden: { type: string, enum: [PENDIENTE, EN_EJECUCION, CULMINADO, CANCELADO, VERIFICADO], example: EN_EJECUCION }
        nro: { type: string, description: "`nro_orden` formateado con ceros a la izquierda + `/gestion`.", example: "000001/2026" }
        tipo_orden: { type: string, enum: [INTERNO, EXTERNO], description: "Derivado: `EXTERNO` si tiene `id_taller`, `INTERNO` si no.", example: INTERNO }
        detalles_count: { type: integer, description: "Sólo en el listado (`GET /ordenes-trabajo`). Cantidad de ítems de detalle registrados.", example: 3 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        vehiculo: { $ref: '#/components/schemas/Vehiculo' }
        solicitudMantenimiento:
          allOf: [{ $ref: '#/components/schemas/SolicitudMantenimiento' }]
          nullable: true
        taller:
          type: object
          nullable: true
          properties:
            id: { type: integer, example: 4 }
            razon_social: { type: string, example: Taller Mecánico Central S.R.L. }
        usuarioEmite:
          allOf: [{ $ref: '#/components/schemas/Usuario' }]
          nullable: true
        usuarioEjecuta:
          allOf: [{ $ref: '#/components/schemas/Usuario' }]
          nullable: true
        detalles:
          type: array
          description: "Ítems del detalle de trabajo. Presente en `GET /ordenes-trabajo/{orden}`."
          items: { $ref: '#/components/schemas/DetalleMantenimiento' }
...
    DetalleMantenimientoStoreRequest:
      type: object
      description: |
        Cuerpo esperado por `POST /ordenes-trabajo/{orden}/detalles`. `id_orden_trabajo` se toma
        de la URL y no debe enviarse.
      required: [id_tipo_mantenimiento, fecha, cantidad]
      properties:
        id_tipo_mantenimiento:
          type: integer
          description: Debe existir en el catálogo `tipo_mantenimiento` y ser de ámbito `taller` (ver `GET /parametros/colecciones` → `ordenes_trabajo.tipos_mantenimiento`).
          example: 3
        id_repuesto:
          type: integer
          nullable: true
          description: Opcional. Debe existir en `repuesto` (ver `GET /parametros/colecciones` → `ordenes_trabajo.repuestos`). Se omite cuando el ítem es sólo mano de obra.
          example: 8
        fecha:
          type: string
          format: date
          example: "2026-08-27"
        horometro:
          type: number
          nullable: true
          minimum: 0
          description: "**Obligatorio si el vehículo de la orden mide por horómetro**; ignorado en caso contrario."
          example: null
        kilometraje:
          type: number
          nullable: true
          minimum: 0
          description: "**Obligatorio si el vehículo de la orden mide por kilometraje**; ignorado en caso contrario."
          example: 85230.5
        cantidad:
          type: integer
          minimum: 1
          example: 2
    DetalleMantenimientoStoreResponse:
      type: object
      properties:
        message: { type: string, example: Detalle de mantenimiento registrado exitosamente. }
        data: { $ref: '#/components/schemas/DetalleMantenimiento' }
    OrdenTrabajoCulminarRequest:
      type: object
      description: |
        Cuerpo esperado por `POST /ordenes-trabajo/{orden}/culminar`. `fecha_ejecucion`,
        `fecha_culminacion` y `estado_orden` se fijan en el servidor y no deben enviarse.
      properties:
        kilometraje_actual:
          type: integer
          nullable: true
          minimum: 0
          description: "Lectura final del odómetro. **Obligatoria si el vehículo de la orden mide por kilometraje.**"
          example: 85300
        horometro_actual:
          type: integer
          nullable: true
          minimum: 0
          description: "Lectura final del horómetro. **Obligatoria si el vehículo de la orden mide por horómetro.**"
          example: null
        observacion:
          type: string
          nullable: true
          maxLength: 500
          example: Se cambió aceite y filtros. Todo en orden.
...